+++ lib/generators/onesie/templates/template.rb
@@ -1,0 +1,1 @@
+# Write the contents of Onesie::Tasks#run here!