+++ lib/onesie/template_reader.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Onesie
+  # Responsible for loading custom template files
+  class TemplateReader
+    def self.read_template(path)
+      File.read(path)
+    end
+  end
+end